requirements.txt
anthropic>=0.83.0
pyyaml>=6.0
